updated unified framework submitted to app stores

+++ lib/features/categories/categories.service.dart
@@ -44,6 +44,8 @@ class CategoriesService extends GetxService with ConsoleMixin {
   }
 
   Future<void> clear() async {
+    if (box?.isOpen == false) return;
+
     if (!boxInitialized) {
       await File(join(LisoPaths.hivePath, '$kHiveBoxCategories.hive')).delete();
       return;
